handle $http error

+++ src/SimplCommerce.Web/Areas/Admin/Controllers/CategoryController.cs
@@ -7,6 +7,7 @@ using SimplCommerce.Web.Areas.Admin.Helpers;
 using SimplCommerce.Web.Areas.Admin.ViewModels;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Authorization;
+using Microsoft.EntityFrameworkCore;
 
 namespace SimplCommerce.Web.Areas.Admin.Controllers
 {
@@ -88,15 +89,20 @@ namespace SimplCommerce.Web.Areas.Admin.Controllers
         [HttpPost]
         public IActionResult Delete(long id)
         {
-            var category = categoryRepository.Query().FirstOrDefault(x => x.Id == id);
+            var category = categoryRepository.Query().Include(x => x.Child).FirstOrDefault(x => x.Id == id);
             if (category == null)
             {
                 return new NotFoundResult();
             }
 
+            if(category.Child.Any(x => !x.IsDeleted))
+            {
+                return BadRequest(new { Error = "Please make sure this category contains no children" });
+            }
+
             categoryService.Delete(category);
 
-            return Json(true);
+            return Ok();
         }
     }
 }
